--- designsystems/mui/components/alert/ErrorAlert.tsx
@@ -1,9 +1,0 @@
-import { Alert } from '@mui/material';
-
-interface IErrorAlertProps {
-  message?: string;
-}
-
-export function ErrorAlert({ message }: IErrorAlertProps) {
-  return <>{message && <Alert severity="error">{message}</Alert>}</>;
-}
