+++ designsystems/mui/components/card/BrothCard.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
   Box,
   Card as MuiCard,
