--- designsystems/mui/features/authorization/PasswordResetForm.tsx
@@ -1,41 +1,0 @@
-import { useForm } from 'react-hook-form';
-import {
-  PasswordResetType,
-  PasswordResetResolver,
-} from '@/schemas/PasswordResetSchema';
-import { Button } from '@mui/material';
-import { MuiTextField } from '../../components/textfield/MuiTextField';
-import BrothLink from '@/components/link/BrothLink';
-
-// TODO: decide password reset functionality in demo or not
-export function PasswordResetForm() {
-  const { handleSubmit, control } = useForm<PasswordResetType>({
-    resolver: PasswordResetResolver,
-    defaultValues: {
-      username: '',
-    },
-  });
-
-  const onSubmit = async () => {};
-
-  return (
-    <form onSubmit={handleSubmit(onSubmit)}>
-      <div>
-        <MuiTextField<PasswordResetType>
-          id="username"
-          name="username"
-          type="email"
-          label="email"
-          control={control}
-          fullWidth
-          margin="normal"
-          helperText={<BrothLink href="/">Forgot your password?</BrothLink>}
-        />
-      </div>
-
-      <Button variant="contained" type="submit" fullWidth>
-        Reset Password
-      </Button>
-    </form>
-  );
-}
